refactor: massive refactor
