+++ uniswap/constants.py
@@ -1,0 +1,39 @@
+ETH_ADDRESS = "0x0000000000000000000000000000000000000000"
+WETH9_ADDRESS = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2"
+
+# see: https://chainid.network/chains/
+_netid_to_name = {
+    1: "mainnet",
+    3: "ropsten",
+    4: "rinkeby",
+    56: "binance",
+    97: "binance_testnet",
+    100: "xdai",
+}
+
+_factory_contract_addresses_v1 = {
+    "mainnet": "0xc0a47dFe034B400B47bDaD5FecDa2621de6c4d95",
+    "ropsten": "0x9c83dCE8CA20E9aAF9D3efc003b2ea62aBC08351",
+    "rinkeby": "0xf5D915570BC477f9B8D6C0E980aA81757A3AaC36",
+    "kovan": "0xD3E51Ef092B2845f10401a0159B2B96e8B6c3D30",
+    "görli": "0x6Ce570d02D73d4c384b46135E87f8C592A8c86dA",
+}
+
+
+# For v2 the address is the same on mainnet, Ropsten, Rinkeby, Görli, and Kovan
+# https://uniswap.org/docs/v2/smart-contracts/factory
+_factory_contract_addresses_v2 = {
+    "mainnet": "0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f",
+    "ropsten": "0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f",
+    "rinkeby": "0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f",
+    "görli": "0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f",
+    "xdai": "0xA818b4F111Ccac7AA31D0BCc0806d64F2E0737D7",
+}
+
+_router_contract_addresses_v2 = {
+    "mainnet": "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D",
+    "ropsten": "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D",
+    "rinkeby": "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D",
+    "görli": "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D",
+    "xdai": "0x1C232F01118CB8B424793ae03F870aa7D0ac7f77",
+}
