+++ uniswap/util.py
@@ -1,0 +1,58 @@
+import os
+import json
+import functools
+from typing import Union
+
+from web3 import Web3
+
+from .types import AddressLike, Address, ENS, Contract
+from .exceptions import InvalidToken
+
+
+def _str_to_addr(s: Union[str, Address]) -> AddressLike:
+    """Idempotent"""
+    if isinstance(s, str):
+        if s.startswith("0x"):
+            return Address(bytes.fromhex(s[2:]))
+        elif s.endswith(".eth"):
+            return ENS(s)
+        else:
+            raise Exception(f"Couldn't convert string '{s}' to AddressLike")
+    else:
+        return s
+
+
+def _addr_to_str(a: AddressLike) -> str:
+    if isinstance(a, bytes):
+        # Address or ChecksumAddress
+        addr: str = Web3.toChecksumAddress("0x" + bytes(a).hex())
+        return addr
+    elif isinstance(a, str):
+        if a.endswith(".eth"):
+            # Address is ENS
+            raise Exception("ENS not supported for this operation")
+        elif a.startswith("0x"):
+            addr = Web3.toChecksumAddress(a)
+            return addr
+
+    raise InvalidToken(a)
+
+
+def _validate_address(a: AddressLike) -> None:
+    assert _addr_to_str(a)
+
+
+def _load_abi(name: str) -> str:
+    path = f"{os.path.dirname(os.path.abspath(__file__))}/assets/"
+    with open(os.path.abspath(path + f"{name}.abi")) as f:
+        abi: str = json.load(f)
+    return abi
+
+
+@functools.lru_cache()
+def _load_contract(w3: Web3, abi_name: str, address: AddressLike) -> Contract:
+    return w3.eth.contract(address=address, abi=_load_abi(abi_name))
+
+
+def _load_contract_erc20(w3: Web3, address: AddressLike) -> Contract:
+    return _load_contract(w3, "erc20", address)
